feat: move modules package to controller's one
