+++ server/server.go
@@ -11,11 +11,11 @@ import (
 	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
 	"github.com/soprasteria/docktor/server/controllers"
-	"github.com/soprasteria/docktor/server/modules/auth"
-	"github.com/soprasteria/docktor/server/modules/daemons"
-	"github.com/soprasteria/docktor/server/modules/groups"
-	"github.com/soprasteria/docktor/server/modules/services"
-	"github.com/soprasteria/docktor/server/modules/users"
+	"github.com/soprasteria/docktor/server/controllers/auth"
+	"github.com/soprasteria/docktor/server/controllers/daemons"
+	"github.com/soprasteria/docktor/server/controllers/groups"
+	"github.com/soprasteria/docktor/server/controllers/services"
+	"github.com/soprasteria/docktor/server/controllers/users"
 	"github.com/soprasteria/docktor/server/storage"
 	"github.com/soprasteria/docktor/server/types"
 	"github.com/spf13/viper"
